gxp-chif-svc: extract UEFI firmware volume parser into uefi_fv
Split the generic UEFI FV/FFS file extraction logic out of
platdef_extract.cpp into its own compilation unit. This makes the
UEFI parser reusable and keeps platdef_extract.cpp focused on
PlatDef-specific decompression and record parsing.

New files:
  - uefi_fv.hpp: EfiGuid struct and extractFfsFile() API
  - uefi_fv.cpp: FV scanning, FFS walking, section handling

The extractPlatDef() function now opens the MTD device and delegates
FV/FFS extraction to extractFfsFile(), then decompresses the result.
No functional change.
