+++ meta-canopy/meta-hpe/meta-common/recipes-hpe/gxp-chif-service/gxp-chif-service/src/platdef_extract.cpp
@@ -1,12 +1,14 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (C) 2026 9elements GmbH
 
-// Extract the APML Platform Definition blob from the host BIOS SPI flash.
-// The blob is stored inside a UEFI firmware volume as an FFS file,
-// identified by well-known GUIDs. This replaces HPE's standalone
-// hpe-platdef-extract utility with an in-process implementation.
+// Parse the HPE Platform Definition (PlatDef) blob and extract I2C
+// segment-to-bus mappings for the CHIF I2C proxy. The raw blob is
+// extracted from the host BIOS SPI flash via the UEFI FV parser
+// (uefi_fv.hpp), then decompressed and walked for I2CEngine records.
 
 #include "platdef_extract.hpp"
+
+#include "uefi_fv.hpp"
 
 #include <phosphor-logging/lg2.hpp>
 
@@ -21,55 +23,8 @@
 namespace chif
 {
 
-// UEFI packed structures for firmware volume parsing
-#pragma pack(push, 1)
-
-struct EfiGuid
-{
-    uint32_t data1;
-    uint16_t data2;
-    uint16_t data3;
-    uint8_t data4[8];
-};
-
-struct EfiFvHeader
-{
-    uint8_t zeroVector[16];
-    EfiGuid fileSystemGuid;
-    uint64_t fvLength;
-    uint32_t signature; // '_FVH' = 0x4856465F
-    uint32_t attributes;
-    uint16_t headerLength;
-    uint16_t checksum;
-    uint16_t extHeaderOffset;
-    uint8_t reserved;
-    uint8_t revision;
-    // BlockMap follows but we don't need it
-};
-
-struct EfiFvExtHeader
-{
-    EfiGuid fvName;
-    uint32_t extHeaderSize;
-};
-
-struct EfiFfsFileHeader
-{
-    EfiGuid name;
-    uint16_t integrityCheck;
-    uint8_t type;
-    uint8_t attributes;
-    uint8_t size[3];
-    uint8_t state;
-};
-
-struct EfiSectionHeader
-{
-    uint8_t size[3];
-    uint8_t type;
-};
-
 // PlatDef bundle and record structures (from HPE openbmc-chif-svc platdef.h)
+#pragma pack(push, 1)
 
 // From HPE upstream platdef.h
 #define PLATDEF_BUNDLE_SIGNATURE "$PlatdefBundle1$"
@@ -178,47 +133,6 @@ static_assert(sizeof(PlatDefI2CMux) == 16,
 static_assert(sizeof(PlatDefI2CEngineFixed) == 48,
               "PlatDefI2CEngineFixed must be 48 bytes");
 
-static constexpr uint32_t efiFvhSignature = 0x4856465F; // '_FVH'
-static constexpr uint8_t efiFvFileTypeFreeform = 0x02;
-static constexpr uint8_t efiFfsLargeFile = 0x01;
-static constexpr uint8_t efiSectionRaw = 0x19;
-static constexpr uint8_t efiSectionGuidDefined = 0x02;
-
-// APML firmware volume GUID: {7EBF5AB8-525E-417C-9B6B-5EF367856954}
-static constexpr EfiGuid apmlFvGuid = {
-    0x7EBF5AB8, 0x525E, 0x417C,
-    {0x9B, 0x6B, 0x5E, 0xF3, 0x67, 0x85, 0x69, 0x54}};
-
-// APML file GUID: {C5F6001C-39B4-43DD-9B9B-6832F1BB4BE9}
-static constexpr EfiGuid apmlFileGuid = {
-    0xC5F6001C, 0x39B4, 0x43DD,
-    {0x9B, 0x9B, 0x68, 0x32, 0xF1, 0xBB, 0x4B, 0xE9}};
-
-// EFI FFS2 file system GUID: {8C8CE578-8A3D-4F1C-9935-896185C32DD3}
-static constexpr EfiGuid ffs2Guid = {
-    0x8C8CE578, 0x8A3D, 0x4F1C,
-    {0x99, 0x35, 0x89, 0x61, 0x85, 0xC3, 0x2D, 0xD3}};
-
-static bool guidEqual(const EfiGuid& a, const EfiGuid& b)
-{
-    return std::memcmp(&a, &b, sizeof(EfiGuid)) == 0;
-}
-
-static bool guidIsAllFf(const EfiGuid& g)
-{
-    static constexpr EfiGuid allFf = {
-        0xFFFFFFFF, 0xFFFF, 0xFFFF,
-        {0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF}};
-    return guidEqual(g, allFf);
-}
-
-static uint32_t ffsFileSize(const EfiFfsFileHeader& hdr)
-{
-    return static_cast<uint32_t>(hdr.size[0]) |
-           (static_cast<uint32_t>(hdr.size[1]) << 8) |
-           (static_cast<uint32_t>(hdr.size[2]) << 16);
-}
-
 // Find the MTD device path for a partition by label
 static std::string findMtdByLabel(const std::string& label)
 {
@@ -251,7 +165,6 @@ static std::string findMtdByLabel(const std::string& label)
         std::getline(nameFile, name);
         if (name == label)
         {
-            // Return /dev/mtdN
             return "/dev/" + entry.path().filename().string();
         }
     }
@@ -307,7 +220,6 @@ static std::vector<uint8_t> decompressPlatDef(const std::vector<uint8_t>& raw)
 
     if (!(tableHdr.flags & tableDataFlagZLib))
     {
-        // Not compressed — return data starting from the table header
         lg2::info("PlatDef: not compressed, returning raw records");
         return std::vector<uint8_t>(raw.begin() + offset, raw.end());
     }
@@ -336,7 +248,6 @@ static std::vector<uint8_t> decompressPlatDef(const std::vector<uint8_t>& raw)
     }
     std::vector<uint8_t> result(sizeof(PlatDefTableData) + decompSize);
 
-    // Copy the table header as-is
     std::memcpy(result.data(), &tableHdr, sizeof(PlatDefTableData));
 
     int rc = ::uncompress(
@@ -354,6 +265,16 @@ static std::vector<uint8_t> decompressPlatDef(const std::vector<uint8_t>& raw)
               "SZ", decompSize, "CNT", tableHdr.recordCount);
     return result;
 }
+
+// APML firmware volume GUID: {7EBF5AB8-525E-417C-9B6B-5EF367856954}
+static constexpr EfiGuid apmlFvGuid = {
+    0x7EBF5AB8, 0x525E, 0x417C,
+    {0x9B, 0x6B, 0x5E, 0xF3, 0x67, 0x85, 0x69, 0x54}};
+
+// APML file GUID: {C5F6001C-39B4-43DD-9B9B-6832F1BB4BE9}
+static constexpr EfiGuid apmlFileGuid = {
+    0xC5F6001C, 0x39B4, 0x43DD,
+    {0x9B, 0x9B, 0x68, 0x32, 0xF1, 0xBB, 0x4B, 0xE9}};
 
 std::vector<uint8_t> extractPlatDef()
 {
@@ -382,153 +303,14 @@ std::vector<uint8_t> extractPlatDef()
     auto romSize = static_cast<size_t>(pos);
     rom.seekg(0);
 
-    // Scan for firmware volumes every 64KB
-    for (size_t fvOffset = 0; fvOffset + sizeof(EfiFvHeader) < romSize;
-         fvOffset += 0x10000)
+    auto raw = extractFfsFile(rom, romSize, apmlFvGuid, apmlFileGuid);
+    if (raw.empty())
     {
-        EfiFvHeader fvHdr{};
-        rom.seekg(static_cast<std::streamoff>(fvOffset));
-        rom.read(reinterpret_cast<char*>(&fvHdr), sizeof(fvHdr));
-        if (!rom.good())
-        {
-            break;
-        }
-
-        if (fvHdr.signature != efiFvhSignature)
-        {
-            continue;
-        }
-
-        if (!guidEqual(fvHdr.fileSystemGuid, ffs2Guid))
-        {
-            continue;
-        }
-
-        if (fvHdr.extHeaderOffset == 0)
-        {
-            continue;
-        }
-
-        // Read extended header to check the volume name GUID
-        EfiFvExtHeader extHdr{};
-        rom.seekg(
-            static_cast<std::streamoff>(fvOffset + fvHdr.extHeaderOffset));
-        rom.read(reinterpret_cast<char*>(&extHdr), sizeof(extHdr));
-        if (!rom.good())
-        {
-            continue;
-        }
-
-        if (!guidEqual(extHdr.fvName, apmlFvGuid))
-        {
-            continue;
-        }
-
-        lg2::info("PlatDef: found APML FV at offset 0x{OFF}",
-                  "OFF", lg2::hex, fvOffset);
-
-        // Walk FFS file entries starting after the extended header
-        size_t fileOffset = fvOffset + fvHdr.extHeaderOffset +
-                            extHdr.extHeaderSize;
-        size_t fvEnd = fvOffset + static_cast<size_t>(fvHdr.fvLength);
-        if (fvEnd > romSize)
-        {
-            fvEnd = romSize;
-        }
-
-        while (fileOffset + sizeof(EfiFfsFileHeader) < fvEnd)
-        {
-            // 8-byte align
-            if (fileOffset & 0x07)
-            {
-                fileOffset = (fileOffset & ~0x07UL) + 0x08;
-            }
-
-            EfiFfsFileHeader fileHdr{};
-            rom.seekg(static_cast<std::streamoff>(fileOffset));
-            rom.read(reinterpret_cast<char*>(&fileHdr), sizeof(fileHdr));
-            if (!rom.good())
-            {
-                break;
-            }
-
-            // End of file list
-            if (guidIsAllFf(fileHdr.name))
-            {
-                break;
-            }
-
-            uint32_t fileSize = ffsFileSize(fileHdr);
-            if (fileSize < sizeof(EfiFfsFileHeader))
-            {
-                break;
-            }
-
-            uint32_t dataSize = fileSize - sizeof(EfiFfsFileHeader);
-            size_t dataOffset = fileOffset + sizeof(EfiFfsFileHeader);
-
-            if (guidEqual(fileHdr.name, apmlFileGuid) &&
-                !(fileHdr.attributes & efiFfsLargeFile))
-            {
-                lg2::info("PlatDef: found APML file, {SZ} bytes at 0x{OFF}",
-                          "SZ", dataSize, "OFF", lg2::hex, dataOffset);
-
-                // Handle section header for FREEFORM files
-                if (fileHdr.type == efiFvFileTypeFreeform &&
-                    dataSize > sizeof(EfiSectionHeader))
-                {
-                    EfiSectionHeader secHdr{};
-                    rom.seekg(static_cast<std::streamoff>(dataOffset));
-                    rom.read(reinterpret_cast<char*>(&secHdr),
-                             sizeof(secHdr));
-                    if (!rom.good())
-                    {
-                        lg2::error("PlatDef: section header read failed");
-                        return {};
-                    }
-
-                    if (secHdr.type == efiSectionRaw)
-                    {
-                        dataOffset += sizeof(EfiSectionHeader);
-                        dataSize -= sizeof(EfiSectionHeader);
-                    }
-                    else if (secHdr.type == efiSectionGuidDefined)
-                    {
-                        constexpr uint32_t guidSecHdrSize = 20;
-                        if (dataSize <= guidSecHdrSize)
-                        {
-                            break;
-                        }
-                        dataOffset += guidSecHdrSize;
-                        dataSize -= guidSecHdrSize;
-                    }
-                }
-
-                // Read the raw APML blob
-                std::vector<uint8_t> raw(dataSize);
-                rom.seekg(static_cast<std::streamoff>(dataOffset));
-                rom.read(reinterpret_cast<char*>(raw.data()),
-                         static_cast<std::streamsize>(dataSize));
-
-                if (!rom.good())
-                {
-                    lg2::error("PlatDef: read failed at offset 0x{OFF}",
-                               "OFF", lg2::hex, dataOffset);
-                    return {};
-                }
-
-                lg2::info("PlatDef: read {SZ} bytes from ROM", "SZ",
-                          raw.size());
-
-                return decompressPlatDef(raw);
-            }
-
-            fileOffset += fileSize;
-        }
+        lg2::info("PlatDef: APML firmware volume not found in ROM");
+        return {};
     }
 
-    lg2::info("PlatDef: APML firmware volume not found in ROM");
-    return {};
+    return decompressPlatDef(raw);
 }
 
 // ---------------------------------------------------------------------------
